Ihux

+++ include/zephyr/pico/button.h
@@ -56,7 +56,7 @@ static inline struct pico_button *
   static int n = 0;  // number of initialized LEDs
   if (cb) n = 0;     // re-init if callback provided
 
-	for (; n < sizeof(but)/sizeof(but[0]); n++) {
+	for (; n < (int)(sizeof(but)/sizeof(but[0])); n++) {
     struct pico_button *p = but + n;
     if (!device_is_ready(p->ds.port)) {
       printk("error %d: BUTTON device not ready\n",-ENODEV);
